PyQtGUI/requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.1
idna==2.8
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyQt5==5.12.1
PyQt5-sip==4.19.15
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
termcolor==1.1.0
twitter==1.18.0
uritemplate==3.0.0
urllib3==1.24.1
